sm6375-common: Configure AAPT
PRODUCT_AAPT_CONFIG:
This variable defines which resource configurations should be generated
and included in the final APK or system image. It primarily affects the
localization (language and region) and resource qualifier options such
as screen density and layout direction.

PRODUCT_AAPT_PREF_CONFIG:
This variable is related to how AAPT selects resource configurations
when packaging an application or system image.

Signed-off-by: Vishnuvardhan Reddy <vvr-rockstar@bmsit.in>