sm6375-common: Enable BUILD_BROKEN_ELF_PREBUILT_PRODUCT_COPY_FILES flag
This flag allows the build system to copy ELF (Executable and Linkable Format)
prebuilt files to the product partition, even if they do not conform to certain
expected standards.
This is useful for including prebuilt binaries that may not fully adhere to the
expected format but are necessary for the device's functionality.

Signed-off-by: Vishnuvardhan Reddy <vvr-rockstar@bmsit.in>